requirements.txt
pymupdf==1.24.9
pillow==10.4.0
pydantic==2.8.2
httpx==0.27.2
tenacity==8.5.0
typer==0.12.3
rich==13.7.1
pandas==2.2.2
numpy==2.0.1
pyyaml==6.0.2
jinja2==3.1.4

openai==1.45.0
google-genai==0.3.0
anthropic==0.34.0
groq==0.12.0
